nlp-pipeline/requirements.txt
# Core ML/DL Libraries
tensorflow==2.13.0
torch==2.0.1
transformers==4.30.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# NLP Libraries
nltk==3.8.1
spacy==3.6.0
gensim==4.3.1
textblob==0.17.1

# API & Deployment
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
redis==4.6.0

# Data Processing
h5py==3.9.0
tqdm==4.65.0
joblib==1.3.1

# Monitoring & Logging
prometheus-client==0.17.1
psutil==5.9.5

# Testing
pytest==7.4.0
pytest-cov==4.1.0
locust==2.15.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0